backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
pymongo==4.5.0
redis==5.0.1
celery==5.3.4
pandas==2.0.3
numpy==1.24.3
MetaTrader5==5.0.43
python-telegram-bot==20.3
talib-binary==0.4.28
matplotlib==3.7.2
mplfinance==0.12.10b0
requests==2.31.0
websockets==12.0
aiohttp==3.8.5
python-socketio==5.9.0
eventlet==0.33.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
pytest==7.4.2
pytest-cov==4.1.0